requirements.txt
ipykernel==6.29.5
numpy==1.26.4
plotly==5.24.1
PyYAML==6.0.2
tqdm==4.67.0